# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Our Empire - Satta Matka Application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
              required: [phone]
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                otp:
                  type: string
              required: [phone, otp]
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                otp:
                  type: string
                newPassword:
                  type: string
              required: [phone, otp, newPassword]
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/profile:
    get:
      operationId: getProfile
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # USER
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin-users]
      summary: Get all users
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin-users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
  /admin/users/{id}/status:
    put:
      operationId: adminUpdateUserStatus
      tags: [admin-users]
      summary: Update user status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, blocked, suspended]
              required: [status]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{id}/balance:
    put:
      operationId: adminUpdateUserBalance
      tags: [admin-users]
      summary: Adjust user balance
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                reason:
                  type: string
              required: [amount, reason]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{id}/activity:
    get:
      operationId: adminGetUserActivity
      tags: [admin-users]
      summary: Get user activity log
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                type: object
                properties:
                  activities:
                    type: array
                    items:
                      $ref: "#/components/schemas/ActivityLog"
                required: [activities]
  /admin/users/{id}/bets:
    get:
      operationId: adminGetUserBets
      tags: [admin-users]
      summary: Get bets for a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bets
          content:
            application/json:
              schema:
                type: object
                properties:
                  bets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Bet"
                required: [bets]

  # ADMIN BETS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required: [message, success]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]
...
    RegisterRequest:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
        referralCode:
          type: string
      required: [fullName, phone, password]
    LoginRequest:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required: [phone, password]
...
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        bankAccountNumber:
          type: string
        ifscCode:
          type: string
        accountHolderName:
          type: string
        upiId:
          type: string
    UserProfile:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        balance:
          type: number
        referralCode:
          type: string
        status:
          type: string
        kycStatus:
          type: string
        createdAt:
          type: string
        bankAccountNumber:
          type: string
        ifscCode:
          type: string
        accountHolderName:
          type: string
        upiId:
          type: string
      required: [id, fullName, phone, balance, referralCode, status, createdAt]
...
    UserDetail:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        balance:
          type: number
        referralCode:
          type: string
        status:
          type: string
        kycStatus:
          type: string
        createdAt:
          type: string
        lastLogin:
          type: string
        totalBets:
          type: integer
        totalDeposits:
          type: number
        totalWithdrawals:
          type: number
      required: [id, fullName, phone, balance, status, createdAt]
...
    Bet:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        marketId:
          type: string
        marketName:
          type: string
        gameType:
          type: string
          enum: [jantri, open, crossing, no_to_no]
        numbers:
          type: array
          items:
            type: object
            properties:
              number:
                type: string
              amount:
                type: number
            required: [number, amount]
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, won, lost, cancelled]
        winAmount:
          type: number
        createdAt:
          type: string
        settledAt:
          type: string
      required: [id, marketId, gameType, numbers, totalAmount, status, createdAt]
...
    PaginatedUsers:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserDetail"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, totalPages]
...
    ActivityLog:
      type: object
      properties:
        id:
          type: string
        action:
          type: string
        details:
          type: object
        ipAddress:
          type: string
        createdAt:
          type: string
        userId:
          type: string
        userName:
          type: string
      required: [id, action, createdAt]